09 cleanArchitecture + provider + getIt + router + state + result + freezed + showSnackBar + showDialog
