+++ lib/main.dart
@@ -1,10 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:image_search_app/data/repository/Image_repository_impl.dart';
 import 'package:image_search_app/di/di_setup.dart';
 import 'package:image_search_app/routes.dart';
-import 'package:image_search_app/ui/image/image_screen.dart';
-import 'package:image_search_app/ui/image/image_view_model.dart';
-import 'package:provider/provider.dart';
