+++ lib/presentation/image/image_state.freezed.dart
